"()[]{}"
this was a tricky problem. at first we have to create a object for brackets pair. first one part will be the key and second part will be the value. after receiving the output ,convert it into a array. then apply for loop. after checking the key we will push it into a temporary array. then check it by value  whether it is exist, in second iteration. for true condition temporary array will be empty.