Given two binary strings a and b, return their sum as a binary string.
at first we have to determine which string is larger . then we use for loop. added the last element of those string as number.  result is a string , if val=2 then we have to add 0 or 1. because Val cant be bigger than 2. so we used modulus operator for it. also if Val=2, then carry=1 . so we used a little trick for carry